+++ packages/ckeditor5-package-generator/lib/index.js
@@ -7,21 +7,23 @@
 
 'use strict';
 
+const { EOL } = require( 'os' );
 const path = require( 'path' );
 const fs = require( 'fs' );
-const { execSync, spawnSync } = require( 'child_process' );
+const { execSync, spawn } = require( 'child_process' );
 
 const { Command } = require( 'commander' );
-const mkdirp = require( 'mkdirp' );
+const chalk = require( 'chalk' );
 const template = require( 'lodash.template' );
 const glob = require( 'glob' );
-const chalk = require( 'chalk' );
+const mkdirp = require( 'mkdirp' );
 
 const packageJson = require( '../package.json' );
 const TEMPLATE_PATH = path.join( __dirname, 'templates' );
 
 const getDependenciesVersions = require( './utils/get-dependencies-versions' );
 const validatePackageName = require( './utils/validate-package-name' );
+const createSpinner = require( './utils/create-spinner' );
 
 // Files that need to be filled with data.
 const TEMPLATES_TO_FILL = [
@@ -31,8 +33,19 @@ const TEMPLATES_TO_FILL = [
 	'README.md'
 ];
 
+// Npm does not publish the `.gitignore` file even if it's somewhere inside the package.
+// Hence, the package generator will create it manually. See: #50.
+const GITIGNORE_ENTRIES = [
+	'build/',
+	'coverage/',
+	'node_modules/',
+	'tmp/',
+	'sample/ckeditor.dist.js',
+	''
+];
+
 new Command( packageJson.name )
-	.argument( '<packageName>', 'name of the package (@scope/ckeditor5-*)' )
+	.argument( '[packageName]', 'name of the package (@scope/ckeditor5-*)' )
 	.option( '-v, --verbose', 'output additional logs', false )
 	.option( '--dev', 'execution of the script in the development mode', () => {
 		// An absolute path to the repository that tracks the package.
@@ -48,16 +61,18 @@ new Command( packageJson.name )
 	.parse( process.argv );
 
 /**
- * @param {String} packageName
+ * @param {String|undefined} packageName
  * @param {CKeditor5PackageGeneratorOptions} options
  */
 async function init( packageName, options ) {
 	// 1. Validate the package name.
 	// 2. Create a directory.
-	// 3. Copy files.
-	// 4. Call npm/yarn install.
-	// 5. Initialize the git repository.
-	// 6. Display an instruction what to do next.
+	// 3. Collecting the latest version of CKEditor 5 dependencies.
+	// 4. Copy files.
+	// 5. Call npm/yarn install.
+	// 6. Initialize the git repository.
+	// 7. Install git hooks.
+	// 8. Display an instruction what to do next.
 	//
 	// * Should we validate Node.js version?
 	// * Should we force using Yarn?
@@ -79,7 +94,7 @@ async function init( packageName, options ) {
 		console.log( chalk.red( validationError ) + '\n' );
 
 		console.log( 'Expected pattern:            ' + chalk.green( '@[scope]/ckeditor5-[feature-name]' ) );
-		console.log( 'The provided package name:   ' + chalk.red( packageName ) );
+		console.log( 'The provided package name:   ' + chalk.red( packageName || '' ) );
 		console.log( 'Allowed characters list:     ' + chalk.blue( '0-9 a-z - . _' ) );
 
 		process.exit( 1 );
@@ -101,6 +116,8 @@ async function init( packageName, options ) {
 	console.log( `📍 Creating the directory "${ chalk.cyan( directoryPath ) }".` );
 	mkdirp.sync( directoryPath );
 
+	// (3.)
+	console.log( '📍 Collecting the latest CKEditor 5 packages versions...' );
 	const packageVersions = getDependenciesVersions( {
 		devMode: options.dev
 	} );
@@ -113,7 +130,7 @@ async function init( packageName, options ) {
 		nodir: true
 	} );
 
-	// (3.)
+	// (4.)
 	console.log( '📍 Copying files...' );
 
 	for ( const templatePath of templatesToCopy ) {
@@ -125,31 +142,43 @@ async function init( packageName, options ) {
 				name: packageName,
 				now: new Date(),
 				program,
-				ckeditor5Version: packageVersions.ckeditor5,
-				devUtilsVersion: packageVersions.devUtils,
-				eslintConfigCkeditor5Version: packageVersions.eslintConfigCkeditor5,
-				stylelintConfigCkeditor5Version: packageVersions.stylelintConfigCkeditor5,
-				packageToolsVersion: packageVersions.packageTools,
-				dllFileName: dllConfiguration.fileName,
-				dllLibrary: dllConfiguration.library
+				packageVersions,
+				dll: dllConfiguration,
+				cliSeparator: program === 'npm' ? '-- ' : ''
 			};
 		}
 
 		copyTemplate( templatePath, directoryPath, data );
 	}
 
-	// (4.)
-	console.log( '📍 Installing dependencies...' );
-	installPackages( directoryPath, {
+	// Create the `.gitignore` file. See #50.
+	fs.writeFileSync( path.join( directoryPath, '.gitignore' ), GITIGNORE_ENTRIES.join( EOL ) );
+
+	// (5.)
+	const installSpinner = createSpinner( 'Installing dependencies... ' + chalk.gray.italic( 'It takes a while.' ), {
+		isDisabled: options.verbose
+	} );
+
+	installSpinner.start();
+
+	await installPackages( directoryPath, {
 		useNpm: options.useNpm,
 		verbose: options.verbose
 	} );
 
-	// (5.)
+	installSpinner.finish();
+
+	// (6.)
 	console.log( '📍 Initializing Git repository...' );
 	initializeGitRepository( directoryPath );
 
-	// (6.)
+	// (7.)
+	console.log( '📍 Installing Git hooks...' );
+	await installGitHooks( directoryPath, {
+		verbose: options.verbose
+	} );
+
+	// (8.)
 	console.log();
 	console.log( chalk.green( 'Done!' ) );
 	console.log();
@@ -189,39 +218,86 @@ function copyTemplate( templateFile, packagePath, data ) {
 }
 
 /**
- * @param {String} directoryPath
+ * @param {String} directoryPath An absolute path to the directory where packages should be installed.
  * @param {Object} options
- * @param {Boolean} options.useNpm
- * @param {Boolean} options.verbose
+ * @param {Boolean} options.useNpm Whether to use `npm` instead of `yarn`.
+ * @param {Boolean} options.verbose Whether to display additional logs.
+ * @returns {Promise}
  */
 function installPackages( directoryPath, options ) {
-	const spawnOptions = {
-		encoding: 'utf8',
-		shell: true,
-		cwd: directoryPath,
-		stderr: 'inherit'
-	};
+	return new Promise( ( resolve, reject ) => {
+		const spawnOptions = {
+			encoding: 'utf8',
+			shell: true,
+			cwd: directoryPath,
+			stderr: 'inherit'
+		};
 
-	if ( options.verbose ) {
-		spawnOptions.stdio = 'inherit';
-	}
+		let installTask;
 
-	if ( options.useNpm ) {
-		const npmArguments = [
-			'install',
-			'--prefix',
-			directoryPath
-		];
+		if ( options.verbose ) {
+			spawnOptions.stdio = 'inherit';
+		}
 
-		spawnSync( 'npm', npmArguments, spawnOptions );
-	} else {
-		const yarnArguments = [
-			'--cwd',
-			directoryPath
-		];
+		if ( options.useNpm ) {
+			const npmArguments = [
+				'install',
+				'--prefix',
+				directoryPath
+			];
 
-		spawnSync( 'yarnpkg', yarnArguments, spawnOptions );
-	}
+			installTask = spawn( 'npm', npmArguments, spawnOptions );
+		} else {
+			const yarnArguments = [
+				'--cwd',
+				directoryPath
+			];
+
+			installTask = spawn( 'yarnpkg', yarnArguments, spawnOptions );
+		}
+
+		installTask.on( 'close', exitCode => {
+			if ( exitCode ) {
+				return reject( new Error( 'Installing dependencies finished with an error.' ) );
+			}
+
+			return resolve();
+		} );
+	} );
+}
+
+/**
+ * @param {String} directoryPath An absolute path to the directory where packages should be installed.
+ * @param {Object} options
+ * @param {Boolean} options.verbose Whether to display additional logs.
+ * @returns {Promise}
+ */
+function installGitHooks( directoryPath, options ) {
+	return new Promise( ( resolve, reject ) => {
+		const spawnOptions = {
+			encoding: 'utf8',
+			shell: true,
+			cwd: directoryPath,
+			stderr: 'inherit'
+		};
+
+		const spawnArguments = [ 'rebuild', 'husky' ];
+
+		if ( options.verbose ) {
+			spawnOptions.stdio = 'inherit';
+		}
+
+		// 'rebuild' was added to yarn in version 2, but we use yarn 1, thus only npm can be used.
+		const rebuildTask = spawn( 'npm', spawnArguments, spawnOptions );
+
+		rebuildTask.on( 'close', exitCode => {
+			if ( exitCode ) {
+				return reject( new Error( 'Rebuilding finished with an error.' ) );
+			}
+
+			return resolve();
+		} );
+	} );
 }
 
 /**