+++ packages/ckeditor5-package-tools/lib/index.js
@@ -16,5 +16,17 @@ module.exports = {
 
 	'dll:build'( options ) {
 		return require( '../lib/tasks/dll-build' )( options );
+	},
+
+	'translations:collect'( options ) {
+		return require( '../lib/tasks/translations-collect' )( options );
+	},
+
+	'translations:upload'( options ) {
+		return require( '../lib/tasks/translations-upload' )( options );
+	},
+
+	'translations:download'( options ) {
+		return require( '../lib/tasks/translations-download' )( options );
 	}
 };