+++ packages/ckeditor5-package-tools/lib/utils/get-webpack-config-dll.js
@@ -8,8 +8,11 @@
 /* eslint-env node */
 
 const path = require( 'path' );
+const fs = require( 'fs' );
+
 const webpack = require( 'webpack' );
 const TerserPlugin = require( 'terser-webpack-plugin' );
+const CKEditorWebpackPlugin = require( '@ckeditor/ckeditor5-dev-webpack-plugin' );
 const { getPostCssConfig } = require( '@ckeditor/ckeditor5-dev-utils' ).styles;
 const getThemePath = require( './get-theme-path' );
 
@@ -27,6 +30,31 @@ module.exports = options => {
 
 	// An absolute path to the manifest file that the `DllReferencePlugin` plugin uses for mapping dependencies.
 	const ckeditor5manifestPath = path.join( options.cwd, 'node_modules', 'ckeditor5', 'build', 'ckeditor5-dll.manifest.json' );
+
+	const webpackPlugins = [
+		new webpack.DllReferencePlugin( {
+			manifest: require( ckeditor5manifestPath ),
+			scope: 'ckeditor5/src',
+			name: 'CKEditor5.dll'
+		} ),
+		new webpack.ProvidePlugin( {
+			process: 'process/browser',
+			Buffer: [ 'buffer', 'Buffer' ]
+		} )
+	];
+
+	// If the package contains localization for the English language, which is the default option for the DLL build,
+	// include the CKEditor 5 Webpack plugin that produces translation files.
+	if ( fs.existsSync( path.join( options.cwd, 'lang', 'translations', 'en.po' ) ) ) {
+		webpackPlugins.push(
+			new CKEditorWebpackPlugin( {
+				language: 'en',
+				additionalLanguages: 'all',
+				sourceFilesPattern: /^src[/\\].+\.js$/,
+				skipPluralFormFunction: true
+			} )
+		);
+	}
 
 	return {
 		mode: 'production',
@@ -56,34 +84,13 @@ module.exports = options => {
 
 		watch: options.watch,
 
-		plugins: [
-			// Uncomment the `CKEditorWebpackPlugin` definition if your plugin contains translations files.
-			// Read more about creating localization files.
-			// See: https://ckeditor.com/docs/ckeditor5/latest/framework/guides/deep-dive/ui/localization.html.
-			// new CKEditorWebpackPlugin( {
-			// 	// UI language. Language codes follow the https://en.wikipedia.org/wiki/ISO_639-1 format.
-			// 	language: 'en',
-			// 	additionalLanguages: 'all',
-			// 	sourceFilesPattern: /^src[/\\].+\.js$/,
-			// 	skipPluralFormFunction: true
-			// } ),
-			new webpack.DllReferencePlugin( {
-				manifest: require( ckeditor5manifestPath ),
-				scope: 'ckeditor5/src',
-				name: 'CKEditor5.dll'
-			} ),
-
-			new webpack.ProvidePlugin( {
-				process: 'process/browser',
-				Buffer: [ 'buffer', 'Buffer' ]
-			} )
-		],
+		plugins: webpackPlugins,
 
 		module: {
 			rules: [
 				{
 					test: /\.svg$/,
-					use: 'raw-loader'
+					use: [ 'raw-loader' ]
 				},
 				{
 					test: /\.css$/,