+++ packages/ckeditor5-package-tools/tests/index.js
@@ -25,13 +25,19 @@ describe( 'lib/index', () => {
 			tasks: {
 				start: sinon.stub(),
 				test: sinon.stub(),
-				'dll:build': sinon.stub()
+				dllBuild: sinon.stub(),
+				translationsCollect: sinon.stub(),
+				translationsDownload: sinon.stub(),
+				translationsUpload: sinon.stub()
 			}
 		};
 
 		mockery.registerMock( '../lib/tasks/test', stubs.tasks.test );
 		mockery.registerMock( '../lib/tasks/start', stubs.tasks.start );
-		mockery.registerMock( '../lib/tasks/dll-build', stubs.tasks[ 'dll:build' ] );
+		mockery.registerMock( '../lib/tasks/dll-build', stubs.tasks.dllBuild );
+		mockery.registerMock( '../lib/tasks/translations-collect', stubs.tasks.translationsCollect );
+		mockery.registerMock( '../lib/tasks/translations-download', stubs.tasks.translationsDownload );
+		mockery.registerMock( '../lib/tasks/translations-upload', stubs.tasks.translationsUpload );
 
 		tasks = require( '../lib' );
 	} );
@@ -91,14 +97,71 @@ describe( 'lib/index', () => {
 		it( 'executes the proper function from the "tasks/" directory', () => {
 			tasks[ 'dll:build' ]();
 
-			expect( stubs.tasks[ 'dll:build' ].calledOnce ).to.equal( true );
+			expect( stubs.tasks.dllBuild.calledOnce ).to.equal( true );
 		} );
 
 		it( 'passes arguments directly to the function', () => {
 			const options = { foo: 1, bar: true };
 			tasks[ 'dll:build' ]( options );
 
-			expect( stubs.tasks[ 'dll:build' ].firstCall.args[ 0 ] ).to.deep.equal( options );
+			expect( stubs.tasks.dllBuild.firstCall.args[ 0 ] ).to.deep.equal( options );
+		} );
+	} );
+
+	describe( '#translations:collect', () => {
+		it( 'is available', () => {
+			expect( tasks[ 'translations:collect' ] ).is.a( 'function' );
+		} );
+
+		it( 'executes the proper function from the "tasks/" directory', () => {
+			tasks[ 'translations:collect' ]();
+
+			expect( stubs.tasks.translationsCollect.calledOnce ).to.equal( true );
+		} );
+
+		it( 'passes arguments directly to the function', () => {
+			const options = { foo: 1, bar: true };
+			tasks[ 'translations:collect' ]( options );
+
+			expect( stubs.tasks.translationsCollect.firstCall.args[ 0 ] ).to.deep.equal( options );
+		} );
+	} );
+
+	describe( '#translations:download', () => {
+		it( 'is available', () => {
+			expect( tasks[ 'translations:download' ] ).is.a( 'function' );
+		} );
+
+		it( 'executes the proper function from the "tasks/" directory', () => {
+			tasks[ 'translations:download' ]();
+
+			expect( stubs.tasks.translationsDownload.calledOnce ).to.equal( true );
+		} );
+
+		it( 'passes arguments directly to the function', () => {
+			const options = { foo: 1, bar: true };
+			tasks[ 'translations:download' ]( options );
+
+			expect( stubs.tasks.translationsDownload.firstCall.args[ 0 ] ).to.deep.equal( options );
+		} );
+	} );
+
+	describe( '#translations:upload', () => {
+		it( 'is available', () => {
+			expect( tasks[ 'translations:upload' ] ).is.a( 'function' );
+		} );
+
+		it( 'executes the proper function from the "tasks/" directory', () => {
+			tasks[ 'translations:upload' ]();
+
+			expect( stubs.tasks.translationsUpload.calledOnce ).to.equal( true );
+		} );
+
+		it( 'passes arguments directly to the function', () => {
+			const options = { foo: 1, bar: true };
+			tasks[ 'translations:upload' ]( options );
+
+			expect( stubs.tasks.translationsUpload.firstCall.args[ 0 ] ).to.deep.equal( options );
 		} );
 	} );
 } );